rename component_fallback module
